use popup as main menu use transparent scrim color
