+++ app/src/main/java/com/henrythasler/cyclemap/Favourites.kt
@@ -11,6 +11,7 @@ import androidx.compose.material3.ModalBottomSheet
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
@@ -95,7 +96,7 @@ fun FavouritesSelectionSheet(
         modifier = Modifier
             .padding(12.dp),
         onDismissRequest = onDismiss,
-//        sheetState = SheetState(true, LocalDensity.current)
+        scrimColor = Color.Transparent,
     ) {
         val padding = 8.dp
         favourites?.forEach { item ->
